Moves ResultsGraph to where the rest of the form stuff is
